Handle mixed numbers.
num::Integral type implements float() method that shall convert
integral values that are within the representable range of f64, and that
can be converted without loss of precision or accuracy. For values
that are outside the range function shall panic.

We figure that the range is: -9007199254740992..=9007199254740992

test cases for the same are added under example/mixed_integer.rs